Reorganize DTF pattern and move it to reference module.
