+++ components/datetime/benches/pattern.rs
@@ -6,7 +6,7 @@ mod fixtures;
 
 use criterion::{criterion_group, criterion_main, Criterion};
 
-use icu_datetime::pattern::Pattern;
+use icu_datetime::pattern::reference::Pattern;
 
 fn pattern_benches(c: &mut Criterion) {
     let patterns: Vec<String> = fixtures::get_patterns_fixture().unwrap().0;
