+++ components/datetime/src/fields/length.rs
@@ -2,10 +2,7 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-use core::{
-    cmp::{Ord, PartialOrd},
-    convert::TryFrom,
-};
+use core::cmp::{Ord, PartialOrd};
 use displaydoc::Display;
 
 #[derive(Display, Debug, PartialEq)]
@@ -31,38 +28,24 @@ pub enum FieldLength {
     Six = 6,
 }
 
-impl From<FieldLength> for u8 {
-    fn from(length: FieldLength) -> Self {
-        match length {
-            FieldLength::One => 1,
-            FieldLength::TwoDigit => 2,
-            FieldLength::Abbreviated => 3,
-            FieldLength::Wide => 4,
-            FieldLength::Narrow => 5,
-            FieldLength::Six => 6,
-        }
+impl FieldLength {
+    pub fn u8_in_range(v: &u8) -> bool {
+        (1..=6).contains(v)
+    }
+
+    pub fn idx(&self) -> u8 {
+        *self as usize as u8
+    }
+
+    pub fn from_idx(idx: u8) -> Result<Self, LengthError> {
+        Ok(match idx {
+            1 => Self::One,
+            2 => Self::TwoDigit,
+            3 => Self::Abbreviated,
+            4 => Self::Wide,
+            5 => Self::Narrow,
+            6 => Self::Six,
+            _ => return Err(LengthError::InvalidLength),
+        })
     }
 }
-
-macro_rules! try_field_length {
-    ($i:ty) => {
-        impl TryFrom<$i> for FieldLength {
-            type Error = LengthError;
-
-            fn try_from(input: $i) -> Result<Self, Self::Error> {
-                Ok(match input {
-                    1 => Self::One,
-                    2 => Self::TwoDigit,
-                    3 => Self::Abbreviated,
-                    4 => Self::Wide,
-                    5 => Self::Narrow,
-                    6 => Self::Six,
-                    _ => return Err(LengthError::InvalidLength),
-                })
-            }
-        }
-    };
-}
-
-try_field_length!(u8);
-try_field_length!(usize);
